pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "glogs"
version = "0.1.0"
requires-python = ">=3.9"

dependencies = [
    "openmeteo-requests",
    "snowflake-connector-python",
    "pandas",
    "requests",
    "dbt-snowflake"
]
